requirements.txt
ultralytics
opencv-python
pyttsx3
pandas
tkinter
Pillow
